Only use native pointers for byref when our hands are tied
We now mimic the C# representation wherever possible, where by-ref
semantics are implicit (e.g. when IReferenceExpression is passed to
by-ref parameter or as value type instance of an invocation).

However, if arithmetic or comparison is done with byrefs, we still
fall back to treating them as unmanaged pointers.

Also fix incorrect comments about __refvalue and __arglist.