Fix failing BinSkim.Driver.FunctionalTests
They were not being run from BuildAndTest.cmd, but they were running
and failing in VS. This fixes the test and enables it on the command
line as well.

The baseline had machine, configuration, and runtime-specific strings
in it. The fix is to strip them from the baseline on generation and
from the actual output during the test run.

Also fix up .gitignore to ignore these files and the .vs/ directory and
remove a baseline file that has no actual counterpart.