Change the output for managed checksum algorithm test where checksum is NOT sha256.
Old code had a hardcoded string of SHA-1 in the message that the HASH was not SHA256.

Now the code will take the actual HASH algorithm and use that value in the string.

Before this was misleading that the user would get a message that their file was SHA-1 and not SHA256 but yet it could have been MD5.