Make more basic changes to underlying code that made assumptions about which source code hashing algorithm was used if it was not Sha256.
Now there is a new Md5 CheckSumAlgorithm enum type.

Added new methods to get the algorithm type based on either a HashType or a Guid.
Either way, a ChecksumAlgorithmType is returned where it has been fully tested and doesn't just assume if it's not Sha256 it must be Sha1.

Also, removed the work MD5 from a message where the message is displayed when 1 or more libraries or binaries are found to NOT use Sha256. This was overstating the precision for what was actually know. What was known is just that Sha256 was not specified, but it stated MD5.
Now it is more generic with MD5 left out because we actually don't know whether it was Sha1 or MD5.