Fix `BA2004.EnableSecureSourceCodeHashing` to report the actual broken hash algorithm (rather than always reporting SHA-1). (#868)
* Change the output for managed checksum algorithm test where checksum is NOT sha256.

Old code had a hardcoded string of SHA-1 in the message that the HASH was not SHA256.

Now the code will take the actual HASH algorithm and use that value in the string.

Before this was misleading that the user would get a message that their file was SHA-1 and not SHA256 but yet it could have been MD5.

* Update ReleaseHistory with the PR # and link

* Missed updating one place in the comment for the RuleResources.BA2004_Error_Managed message.

* Make more basic changes to underlying code that made assumptions about which source code hashing algorithm was used if it was not Sha256.

Now there is a new Md5 CheckSumAlgorithm enum type.

Added new methods to get the algorithm type based on either a HashType or a Guid.
Either way, a ChecksumAlgorithmType is returned where it has been fully tested and doesn't just assume if it's not Sha256 it must be Sha1.

Also, removed the work MD5 from a message where the message is displayed when 1 or more libraries or binaries are found to NOT use Sha256. This was overstating the precision for what was actually know. What was known is just that Sha256 was not specified, but it stated MD5.
Now it is more generic with MD5 left out because we actually don't know whether it was Sha1 or MD5.

* Update ReleaseHistory to reflect correct Pull Request and to group this entry with the other Bugs

* Update to using switch statement

* Updating baseline

* Update BinSkimRules.md file after build

* Fix version constants from 4.0.0 to 4.1.0

* Update release history to have a much better title.

---------

Co-authored-by: Craig Schlaman <v-crschlaman@microsoft.com>